Move VirtualNode#createSubnode
